Support pushdown physical sort operator to speedup SortMergeJoin
