+++ opensearch/src/main/java/org/opensearch/sql/opensearch/planner/physical/OpenSearchSortIndexScanRule.java
@@ -8,9 +8,9 @@ package org.opensearch.sql.opensearch.planner.physical;
 import java.util.function.Predicate;
 import org.apache.calcite.plan.RelOptRuleCall;
 import org.apache.calcite.plan.RelRule;
-import org.apache.calcite.rel.logical.LogicalSort;
+import org.apache.calcite.rel.core.Sort;
 import org.immutables.value.Value;
-import org.opensearch.sql.opensearch.storage.scan.CalciteLogicalIndexScan;
+import org.opensearch.sql.opensearch.storage.scan.AbstractCalciteIndexScan;
 
 @Value.Enclosing
 public class OpenSearchSortIndexScanRule extends RelRule<OpenSearchSortIndexScanRule.Config> {
@@ -21,11 +21,11 @@ public class OpenSearchSortIndexScanRule extends RelRule<OpenSearchSortIndexScan
 
   @Override
   public void onMatch(RelOptRuleCall call) {
-    final LogicalSort sort = call.rel(0);
-    final CalciteLogicalIndexScan scan = call.rel(1);
+    final Sort sort = call.rel(0);
+    final AbstractCalciteIndexScan scan = call.rel(1);
 
     var collations = sort.collation.getFieldCollations();
-    CalciteLogicalIndexScan newScan = scan.pushDownSort(collations);
+    AbstractCalciteIndexScan newScan = scan.pushDownSort(collations);
     if (newScan != null) {
       call.transformTo(newScan);
     }
@@ -39,11 +39,11 @@ public class OpenSearchSortIndexScanRule extends RelRule<OpenSearchSortIndexScan
             .build()
             .withOperandSupplier(
                 b0 ->
-                    b0.operand(LogicalSort.class)
+                    b0.operand(Sort.class)
                         .predicate(OpenSearchIndexScanRule::sortByFieldsOnly)
                         .oneInput(
                             b1 ->
-                                b1.operand(CalciteLogicalIndexScan.class)
+                                b1.operand(AbstractCalciteIndexScan.class)
                                     // Skip the rule if a limit has already been pushed down
                                     // because pushing down a sort after a limit will be treated
                                     // as sort-then-limit by OpenSearch DSL.
