+++ opensearch/src/main/java/org/opensearch/sql/opensearch/storage/scan/CalciteLogicalIndexScan.java
@@ -6,12 +6,10 @@
 package org.opensearch.sql.opensearch.storage.scan;
 
 import com.google.common.collect.ImmutableList;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import lombok.Getter;
 import org.apache.calcite.plan.Convention;
 import org.apache.calcite.plan.RelOptCluster;
@@ -26,7 +24,6 @@ import org.apache.calcite.rel.core.Aggregate;
 import org.apache.calcite.rel.core.Filter;
 import org.apache.calcite.rel.core.Project;
 import org.apache.calcite.rel.hint.RelHint;
-import org.apache.calcite.rel.logical.LogicalAggregate;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.rel.type.RelDataTypeFactory;
 import org.apache.calcite.rel.type.RelDataTypeField;
@@ -35,15 +32,10 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.search.aggregations.AggregationBuilder;
-import org.opensearch.search.sort.ScoreSortBuilder;
-import org.opensearch.search.sort.SortBuilder;
-import org.opensearch.search.sort.SortBuilders;
-import org.opensearch.search.sort.SortOrder;
 import org.opensearch.sql.calcite.utils.OpenSearchTypeFactory;
 import org.opensearch.sql.common.setting.Settings;
 import org.opensearch.sql.data.type.ExprType;
 import org.opensearch.sql.opensearch.data.type.OpenSearchDataType;
-import org.opensearch.sql.opensearch.data.type.OpenSearchTextType;
 import org.opensearch.sql.opensearch.planner.physical.EnumerableIndexScanRule;
 import org.opensearch.sql.opensearch.planner.physical.OpenSearchIndexRules;
 import org.opensearch.sql.opensearch.request.AggregateAnalyzer;
@@ -77,6 +69,19 @@ public class CalciteLogicalIndexScan extends AbstractCalciteIndexScan {
       RelDataType schema,
       PushDownContext pushDownContext) {
     super(cluster, traitSet, hints, table, osIndex, schema, pushDownContext);
+  }
+
+  @Override
+  protected AbstractCalciteIndexScan buildScan(
+      RelOptCluster cluster,
+      RelTraitSet traitSet,
+      List<RelHint> hints,
+      RelOptTable table,
+      OpenSearchIndex osIndex,
+      RelDataType schema,
+      PushDownContext pushDownContext) {
+    return new CalciteLogicalIndexScan(
+        cluster, traitSet, hints, table, osIndex, schema, pushDownContext);
   }
 
   public CalciteLogicalIndexScan copyWithNewSchema(RelDataType schema) {
@@ -245,126 +250,5 @@ public class CalciteLogicalIndexScan extends AbstractCalciteIndexScan {
       }
     }
     return null;
-  }
-
-  public CalciteLogicalIndexScan pushDownSort(List<RelFieldCollation> collations) {
-    try {
-      List<String> collationNames = getCollationNames(collations);
-      if (getPushDownContext().isAggregatePushed() && hasAggregatorInSortBy(collationNames)) {
-        // If aggregation is pushed down, we cannot push down sorts where its by fields contain
-        // aggregators.
-        return null;
-      }
-
-      // Propagate the sort to the new scan
-      RelTraitSet traitsWithCollations = getTraitSet().plus(RelCollations.of(collations));
-      CalciteLogicalIndexScan newScan =
-          new CalciteLogicalIndexScan(
-              getCluster(),
-              traitsWithCollations,
-              hints,
-              table,
-              osIndex,
-              getRowType(),
-              // Existing collations are overridden (discarded) by the new collations,
-              cloneWithoutSort(pushDownContext));
-
-      List<SortBuilder<?>> builders = new ArrayList<>();
-      for (RelFieldCollation collation : collations) {
-        int index = collation.getFieldIndex();
-        String fieldName = this.getRowType().getFieldNames().get(index);
-        RelFieldCollation.Direction direction = collation.getDirection();
-        RelFieldCollation.NullDirection nullDirection = collation.nullDirection;
-        // Default sort order is ASCENDING
-        SortOrder order =
-            RelFieldCollation.Direction.DESCENDING.equals(direction)
-                ? SortOrder.DESC
-                : SortOrder.ASC;
-        // TODO: support script sort and distance sort
-        SortBuilder<?> sortBuilder;
-        if (ScoreSortBuilder.NAME.equals(fieldName)) {
-          sortBuilder = SortBuilders.scoreSort();
-        } else {
-          String missing =
-              switch (nullDirection) {
-                case FIRST -> "_first";
-                case LAST -> "_last";
-                default -> null;
-              };
-          // Keyword field is optimized for sorting in OpenSearch
-          String fieldNameKeyword =
-              OpenSearchTextType.convertTextToKeyword(
-                  fieldName, osIndex.getFieldTypes().get(fieldName));
-          sortBuilder = SortBuilders.fieldSort(fieldNameKeyword).missing(missing);
-        }
-        builders.add(sortBuilder.order(order));
-      }
-      newScan.pushDownContext.add(
-          PushDownAction.of(
-              PushDownType.SORT,
-              builders.toString(),
-              requestBuilder -> requestBuilder.pushDownSort(builders)));
-      return newScan;
-    } catch (Exception e) {
-      if (LOG.isDebugEnabled()) {
-        LOG.debug("Cannot pushdown the sort {}", getCollationNames(collations), e);
-      } else {
-        LOG.info("Cannot pushdown the sort {}, ", getCollationNames(collations));
-      }
-    }
-    return null;
-  }
-
-  private List<String> getCollationNames(List<RelFieldCollation> collations) {
-    return collations.stream()
-        .map(collation -> getRowType().getFieldNames().get(collation.getFieldIndex()))
-        .toList();
-  }
-
-  /**
-   * Check if the sort by collations contains any aggregators that are pushed down. E.g. In `stats
-   * avg(age) as avg_age by state | sort avg_age`, the sort clause has `avg_age` which is an
-   * aggregator. The function will return true in this case.
-   *
-   * @param collations List of collation names to check against aggregators.
-   * @return True if any collation name matches an aggregator output, false otherwise.
-   */
-  private boolean hasAggregatorInSortBy(List<String> collations) {
-    Stream<LogicalAggregate> aggregates =
-        pushDownContext.stream()
-            .filter(action -> action.type() == PushDownType.AGGREGATION)
-            .map(action -> ((LogicalAggregate) action.digest()));
-    return aggregates
-        .map(aggregate -> isAnyCollationNameInAggregateOutput(aggregate, collations))
-        .reduce(false, Boolean::logicalOr);
-  }
-
-  private static boolean isAnyCollationNameInAggregateOutput(
-      LogicalAggregate aggregate, List<String> collations) {
-    List<String> fieldNames = aggregate.getRowType().getFieldNames();
-    // The output fields of the aggregate are in the format of
-    // [...grouping fields, ...aggregator fields], so we set an offset to skip
-    // the grouping fields.
-    int groupOffset = aggregate.getGroupSet().cardinality();
-    List<String> fieldsWithoutGrouping = fieldNames.subList(groupOffset, fieldNames.size());
-    return collations.stream()
-        .map(fieldsWithoutGrouping::contains)
-        .reduce(false, Boolean::logicalOr);
-  }
-
-  /**
-   * Create a new {@link PushDownContext} without the collation action.
-   *
-   * @param pushDownContext The original push-down context.
-   * @return A new push-down context without the collation action.
-   */
-  private static PushDownContext cloneWithoutSort(PushDownContext pushDownContext) {
-    PushDownContext newContext = new PushDownContext();
-    for (PushDownAction action : pushDownContext) {
-      if (action.type() != PushDownType.SORT) {
-        newContext.add(action);
-      }
-    }
-    return newContext;
   }
 }
